+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # relysjs
 
+## 0.20.0
+
+### Minor Changes
+
+- minor:
+
+      static_middleware_: + config?:static_middleware__config_T
+      	+ headers_(url_path:string, content_type:string, path:string)=>Record<string, string>
+      static_middleware__config_T
+
 ## 0.19.0
 
 ### Minor Changes